Refactor interfaces, dbt readers and Metabase client
