+++ tests/__init__.py
@@ -1,2 +1,2 @@
-from .test_dbt_parsers import *
+from .test_dbt import *
 from .test_metabase import *
